[REFACTOR] 패키징 변경 universityExamRecord -> examRecord
